requirements.txt
httpx==0.27.0
PyGObject==3.48.2
python-dotenv==1.0.1
